requirements.txt
requests
beautifulsoup4
pandas
selenium
openai
python-dotenv
tiktoken
streamlit
plotly 